find static secret index
Unfortunately in the 4 steps, this algorithm seems to have a strange oddity
where a strong bias is exhibited. A test of independence using the Pearson
Chi-Squared distribution returns a value of 1.9E-8, showing that the generator
most certainly does not show independence between sequential observations.

However, this is due to step dynamically discovering the output of the RNG
based on the layout of the cards in the final step. Instead, by changing this
final step to discover a static secret index that is used for every round, the
bias is eliminated, and the observations fall within the left and right tails
of the Chi-square distribution.

Thus, at the beginning of step 4, after the Talons have been populated, then
gathered, the index is set by reading the top card from the deck. If the top
card is the Ace of Clubs, or the value of "1", then the top card is used for
reading the output of the generator. If the top card is the King of Spades, or
the value of "52", then the bottom card is used for reading the output of the
generator. Et cetera.